+++ tests/test_boxes.py
@@ -1,4 +1,5 @@
 import kwimage
+import pytest
 
 
 def _is_torch_available():
@@ -42,3 +43,17 @@ def test_boxes_clip_torch():
     boxes = kwimage.Boxes.random(10, rng=0).toformat('ltrb').scale(20)
     boxes = boxes.tensor()
     _box_clip_test(boxes)
+
+
+def test_boxes_empty_draw():
+    try:
+        import kwplot
+    except ImportError:
+        pytest.skip('requires kwplot')
+
+    import kwimage
+    self = kwimage.Boxes.random(num=0, format='ltrb')
+    kwplot.figure(fnum=1, doclf=True)
+    #kwplot.imshow(image)
+    # xdoctest: +REQUIRES(--show)
+    self.draw(color='blue', setlim=1.2)