+++ CHANGELOG.md
@@ -4,7 +4,32 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.25 - Unreleased
+## Version 0.10.0 - Unreleased
+
+### Added
+
+* Support for numpy 2.0 on Python 3.9+
+* Add "opacity" argument to `kwimage.Color.adjust`.
+* Add `__json__` to `kwimage.Boxes`
+* Allow user to specify backend order in `load_image_shape`.
+* Add the superstar test image
+* `kwimage.Color.nearest_named` for finding the approximate name of a color.
+
+### Removed
+* Removed support for Python 3.6 and 3.7
+
+
+### Changed
+* Demodata is now more robust to networking outages. If no mirror can be
+  reached, it generates random data to replace the real image.
+
+### Fixed
+* Fix issue with Boxes.draw when Boxes is empty
+* Allow heatmap.draw to do something if classes are not defined.
+
+
+
+## Version 0.9.25 - Released 2024-03-19
 
 ### Fixed